+++ src/main/kotlin/com/fobgochod/git/commit/domain/TypeRow.kt
@@ -6,11 +6,11 @@ import com.fobgochod.git.commit.domain.option.CommitType
  * commit type row
  *
  * @author fobgochod
- * @since 2022/12/11 23:16
+ * @date 2022/12/11 23:16
  */
 data class TypeRow(
-        var name: String = "",
-        var description: String = ""
+    var name: String = "",
+    var description: String = ""
 ) {
 
     constructor(commitType: CommitType) : this() {